Lab5/Kovalenko_AI/requirements.txt
google-genai==1.75.0
python-dotenv==1.0.1